requirements.txt
amqp==5.1.1
asgiref==3.4.1
billiard==4.1.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.8
django-ckeditor==6.2.0
django-js-asset==1.2.2
kombu==5.3.1
Pillow==10.0.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
